fix method: saveWayForPayModule

+++ src/main/java/com/example/learnenglish/controllers/restConrollers/VocabularyPageRestController.java
@@ -10,6 +10,7 @@ package com.example.learnenglish.controllers.restConrollers;
 import com.example.learnenglish.model.VocabularyPage;
 import com.example.learnenglish.model.Image;
 import com.example.learnenglish.responsemessage.CustomResponseMessage;
+import com.example.learnenglish.responsemessage.Message;
 import com.example.learnenglish.service.FileStorageService;
 import com.example.learnenglish.service.VocabularyPageService;
 import lombok.Data;
@@ -38,24 +39,29 @@ public class VocabularyPageRestController {
     @PostMapping("/admin/vocabulary-page/save")
     @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     public ResponseEntity<CustomResponseMessage> vocabularyPageSave(@RequestPart(value = "image", required = false) MultipartFile imageFile,
-                                                                  @RequestPart(value = "vocabularyPage") VocabularyPage vocabularyPage,
-                                                                  Principal principal) throws RuntimeException, IOException {
+                                                                    @RequestPart(value = "vocabularyPage") VocabularyPage vocabularyPage,
+                                                                    Principal principal) throws RuntimeException, IOException {
         if (principal != null) {
+            if(vocabularyPage.getWord() == null || vocabularyPage.getWord().getId() == null) {
+                return ResponseEntity.ok(new CustomResponseMessage(Message.ERROR_REQUIRED_FIELD));
+            }
             try {
                 VocabularyPage vocabularyPageDB = vocabularyPageService.getVocabularyPage(vocabularyPage.getId());
                 vocabularyPage.setImage(new Image());
                 if (imageFile != null) {
                     vocabularyPage.getImage().setImageName(fileStorageService.storeFile(imageFile, wordStoreName, vocabularyPage.getName()));
-                    if (vocabularyPageDB.getImage().getImageName() != null )
+                    if (vocabularyPageDB.getImage().getImageName() != null)
                         fileStorageService.deleteFileFromStorage(vocabularyPageDB.getImage().getImageName(), wordStoreName);
                 }
                 return ResponseEntity.ok(vocabularyPageService.saveVocabularyPage(vocabularyPageDB, vocabularyPage));
 
             } catch (RuntimeException e) {
+                if (vocabularyPageService.existVocabularyPageByName(vocabularyPage.getWord().getName())) {
+                    return ResponseEntity.ok(new CustomResponseMessage(Message.ERROR_DUPLICATE_TEXT));
+                }
                 Image image = new Image();
                 if (imageFile != null)
                     image.setImageName(fileStorageService.storeFile(imageFile, wordStoreName, vocabularyPage.getName()));
-
                 vocabularyPage.setImage(image);
                 return ResponseEntity.ok(vocabularyPageService.saveNewVocabularyPage(vocabularyPage));
             }
